requirements.txt
numpy>=1.24.0
pyyaml>=6.0
websockets>=12.0
rich>=13.0
